+++ app/Controllers/Admin/Login.php
@@ -2,6 +2,7 @@
 
 namespace App\Controllers\Admin;
 use App\Controllers\BaseController;
+use App\Libraries\Hash;
 
 class Login extends BaseController
 {
@@ -11,18 +12,56 @@ class Login extends BaseController
 		$usersModel  = new \App\Models\UsersModel();
         $data = [
 			'titlepage' => 'Login',
-			'view' => 'admin/users/index',
-			'css' => [
-				'DataTables' => 'datatable/datatables.css',
-				'Toastr' => 'toastr/toastr.min.css',
-			],
-			'scripts' => [
-				'DataTables' => 'datatable/datatables.js',
-				'DataTables Default' => 'datatable/default-datatable.js',
-				'Toastr' => 'toastr/toastr.min.js',
-			],
 		];
 		return view('admin/login', $data);
 	}
+	public function check(){
+		helper('form');
+        $validation = $this->validate([
+            'email' => [
+                'rules' => 'required|valid_email|is_not_unique[users.email]',
+                'errors' => [
+                    'required' => 'Digite seu email',
+                    'is_not_unique' => 'Email não registrado',
+                    'valid_email' => 'Digite um email válido',
+                ],
+            ],
+            'password' => [
+                'rules' => 'required|min_length[5]|max_length[12]',
+                'errors' => [
+                    'required' => 'Digite sua senha', 
+                    'min_length' => 'A senha deve ter pelo menos 5 caracteres',
+                    'max_length' => 'A senha não deve ter mais de 12 caracteres',
+                ],                
+            ],
+        ]);
+        if(!$validation){
+			$data=[
+				'titlepage' => 'Login',
+				'validation'=>$this->validator,
+			];
+            return view('admin/login',$data);
+        }else{
+            $email = $this->request->getpost('email');
+            $password = $this->request->getpost('password');
+            $usersModel = new \App\Models\UsersModel();
+            $user_info = $usersModel->where('email',$email)->first();
+            $check_password = Hash::check($password, $user_info['password']);
+            if(!$check_password){
+                session()->setFlashdata('fail','Verifique seu usuário e senha');                
+                return redirect()->to('/admin/login')->withInput();
+            }else{
+                $user_id = $user_info['id'];
+                session()->set('loggedUser', $user_id);
+                return redirect()->to('/admin');
+            }
+        }
+    }
+    public function logout() {
+        if(session()->has('loggedUser')){
+            session()->remove('loggedUser');
+            return redirect()->to('/admin/login?access=out')->with('fail','Você saiu!');
+        }
+    }
 	
 }