+++ app/Views/admin/template.php
@@ -241,6 +241,14 @@
 									</li>
 								</ul>
 							</li>
+							<li class="nav-item">
+								<a href="<?= base_url('admin/login/logout') ?>" class="nav-link">
+									<i class="nav-icon fas fa-sign-out-alt"></i>
+									<p>
+										Sair
+									</p>
+								</a>
+							</li>
 						</ul>
 					</nav>
 					<!-- /.sidebar-menu -->